requirements.txt
# Core Deep Learning Frameworks (Apple Silicon optimized)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLOv8 for Object Detection & Segmentation
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Deep Learning Architectures
segmentation-models-pytorch>=0.3.3
efficientnet-pytorch>=0.7.1
timm>=0.9.0  # PyTorch Image Models

# Scientific Computing
numpy>=1.24.0,<3.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Explainability & Visualization
grad-cam>=1.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# API & Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Handling
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress Bars & Logging
tqdm>=4.65.0
loguru>=0.7.0

# Image Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0  # For testing FastAPI

# Utilities
python-json-logger>=2.0.7
aiofiles>=23.1.0
